% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_parameters.R
\name{make_parameters}
\alias{make_parameters}
\title{Calculate parameter inputs for TMB}
\usage{
make_parameters(Version, DataList, Rank = "Expanded")
}
\arguments{
\item{Version}{a version number (see example for current default).}

\item{DataList}{list outputted from \code{EOFR::make_data}}
}
\description{
\code{Param_Fn} generates the \code{parameters} input for \code{TMB::MakeADFun}
}
